Trait minilzo::OutputBuffer

pub unsafe trait OutputBuffer {
    // Required methods
    fn get_ptr(&mut self) -> *mut MaybeUninit<u8>;
    fn get_size(&self) -> u64;
}

Safety

The as_ptr function MUST return a valid, mutable pointer to potentially uninitialized memory.

The len function MUST return the length of the section of uninitialized memory returned by as_ptr, in bytes.

Required Methods

fn get_ptr(&mut self) -> *mut MaybeUninit<u8>

Get the ptr.

fn get_size(&self) -> u64

Get the length of the buffer.